fix: split ssg-generate.ts to avoid jiti JSX parse error
jiti (rsbuild's config loader) cannot parse JSX in .tsx files.
Extract generateSSGHTML into a plain .ts file so rsbuild.config.ts
can import it without triggering the parser error.

Also add SSG usage documentation to README.
